add some comment and create a QT Creator project
