+++ HomeWork2.pro
@@ -1,0 +1,12 @@
+TEMPLATE = app
+CONFIG += console c++11
+CONFIG -= app_bundle
+CONFIG -= qt
+
+SOURCES += \
+        header.cpp \
+        main.cpp
+
+HEADERS += \
+    header.h \
+    header_cout.h
