+++ header_cout.h
@@ -1,27 +1,35 @@
-void find_min_max(int arr[]);
+#ifndef HEADER_COUT_H
+#define HEADER_COUT_H
 
-int sort_mode(){
-    int num_e = 10;
-    int arr[num_e];
-    int result[2];
-    cout << "Enter 10 numbers (press Enter afterwards): " << "\n";
-    for (int i = 0; i < num_e; ++i) {
-        cin >> arr[i];
-    }
-    find_min_max(arr);
-}
+// Same remark as other file.
+// moreover don't place the cout instruction  in a function. Place it in the main function.
 
-void find_min_max(int arr[]) {
-    int num_e = 10;
-    int min_v = arr[0];
-    int max_v = min_v;
-    for (int i=0; i<num_e; i++){
-        if (arr[i] > max_v) {
-            max_v = arr[i];
-        } else if (arr[i] < min_v) {
-            min_v = arr[i];
-        }
-    }
-    cout << "\nMin value: " << min_v;
-    cout << "\nMax value: " << max_v;
-}
+//void find_min_max(int arr[]);
+
+//int sort_mode(){
+//    int num_e = 10;
+//    int arr[num_e];
+//    int result[2];
+//    cout << "Enter 10 numbers (press Enter afterwards): " << "\n";
+//    for (int i = 0; i < num_e; ++i) {
+//        cin >> arr[i];
+//    }
+//    find_min_max(arr);
+//}
+
+//void find_min_max(int arr[]) {
+//    int num_e = 10;
+//    int min_v = arr[0];
+//    int max_v = min_v;
+//    for (int i=0; i<num_e; i++){
+//        if (arr[i] > max_v) {
+//            max_v = arr[i];
+//        } else if (arr[i] < min_v) {
+//            min_v = arr[i];
+//        }
+//    }
+//    cout << "\nMin value: " << min_v;
+//    cout << "\nMax value: " << max_v;
+//}
+
+#endif // HEADER_COUT_H
