+++ main.cpp
@@ -1,0 +1,13 @@
+#include <iostream>
+
+//using namespace std; --> don't use this instruction. We usually explicitely the namespace that we choose.
+
+
+#include "header.h"
+// #include "header_cout.h"
+
+int main(){
+    std::cout << "....... Exercise 2 .......\n\n";
+    sort_mode();
+    return 0;
+}
